Internal change
PiperOrigin-RevId: 329303148

+++ src/ans_test.cc
@@ -41,7 +41,7 @@ TEST(ANSTest, TestRoundtrip) {
   }
 
   BitWriter writer;
-  std::vector<float> unused_bits_per_ctx;
+  std::vector<double> unused_bits_per_ctx;
   ANSEncode(data, kNumContexts, &writer, &unused_bits_per_ctx);
 
   std::vector<uint8_t> encoded = std::move(writer).GetData();
